requirements.txt
streamlit>=1.30.0
Pillow>=9.0.0
fpdf>=1.7.2
python-docx>=0.8.11
